L2

+++ STPP_Project/Controllers/AccommodationAdsController.cs
@@ -2,6 +2,11 @@
 using STPP_Project.Data.Entities;
 using STPP_Project.Data.Dtos.AccommodationAds;
 using STPP_Project.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
+using System.Data;
+using STPP_Project.Auth.Model;
+using System.Security.Claims;
+using Microsoft.IdentityModel.JsonWebTokens;
 
 namespace STPP_Project.Controllers
 {
@@ -11,10 +16,12 @@ namespace STPP_Project.Controllers
     {
         private readonly IAccommodationAdsRepository _accommodationAdsRepository;
         private readonly ICitiesRepository _citiesRepository;
-        public AccommodationAdsController(IAccommodationAdsRepository accommodationAdsRepository, ICitiesRepository citiesRepository)
+        private readonly IAuthorizationService _authorizationService;
+        public AccommodationAdsController(IAccommodationAdsRepository accommodationAdsRepository, ICitiesRepository citiesRepository, IAuthorizationService authorizationService)
         {
             _accommodationAdsRepository = accommodationAdsRepository;
             _citiesRepository = citiesRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -45,6 +52,7 @@ namespace STPP_Project.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = ProjectRoles.RegisteredUser + ", " + ProjectRoles.Admin)]
         public async Task<ActionResult<AccommodationAdDto>> Create(int cityId, CreateAccommodationAdDto createAccommodationAdDto)
         {
             var city = await _citiesRepository.GetAsync(cityId);
@@ -58,7 +66,8 @@ namespace STPP_Project.Controllers
             {
                 Price = createAccommodationAdDto.Price,
                 BedCount = createAccommodationAdDto.BedCount,
-                City = city
+                City = city,
+                UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)
             };
 
             await _accommodationAdsRepository.CreateAsync(accommodationAd);
@@ -68,6 +77,7 @@ namespace STPP_Project.Controllers
 
         [HttpPut]
         [Route("{accommodationAdId}")]
+        [Authorize(Roles = ProjectRoles.RegisteredUser + ", " + ProjectRoles.Admin)]
         public async Task<ActionResult<AccommodationAdDto>> Update(int cityId, int accommodationAdId, UpdateAccommodationAdDto updateAccommodationDto)
         {
             var city = await _citiesRepository.GetAsync(cityId);
@@ -82,6 +92,12 @@ namespace STPP_Project.Controllers
                 return NotFound($"Couldn't find an ad with ID: {accommodationAdId}");
             }
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, accommodationAd, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             accommodationAd.Price = updateAccommodationDto.Price;
             await _accommodationAdsRepository.UpdateAsync(accommodationAd);
 
@@ -90,6 +106,7 @@ namespace STPP_Project.Controllers
 
         [HttpDelete]
         [Route("{accommodationAdId}")]
+        [Authorize(Roles = ProjectRoles.RegisteredUser + ", " + ProjectRoles.Admin)]
         public async Task<ActionResult> Remove(int cityId, int accommodationAdId)
         {
             var city = await _citiesRepository.GetAsync(cityId);
@@ -103,6 +120,12 @@ namespace STPP_Project.Controllers
             if (accommodationAd == null)
             {
                 return NotFound($"Couldn't find an ad with ID: {accommodationAdId}");
+            }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, accommodationAd, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
             }
 
             await _accommodationAdsRepository.DeleteAsync(accommodationAd);
